made buffer short

+++ Drive/drive_main/drive_ofs.hpp
@@ -48,7 +48,7 @@
 
 #define ADNS3080_PRODUCT_ID_VAL        0x17
 
-int dy_buffer[5] = {0, 0, 0, 0, 0};
+short int dy_buffer[5] = {0, 0, 0, 0, 0};
 struct MD{
  byte motion;
  char dx, dy;
